Role-playing game
Raja Mantri Chor Sipahi (transl.  King, Minister, thief and soldier) is a type of role-playing game that is a popular pastime in India. It involves four players—each player takes up the role of either the king, Minister, thief, or soldier— and the Minister (Mantri) has to guess the identity of the thief. Along with other such games, its popularity has decreased in the 21st century.